A bit of cleanup, but most importantly added Stats() to both the Pool and the Json pool.
Stats() returns a map[string]int64 which represents stats about the pool. Right
now is tracks 2 stats: misses (the number of times an item had to be created
because the pool was empty) and max (the maximum amount of bytes written
to an item). Together, these should provide some helpful information about
pool and item sizing.

The other important stat, which isn't being reported, would be the minimum
size of the pool...that is, just how empty it ever gets.

Calling Stats() resets the counters to 0. Stats isn't thread-safe. The map
which is reused on subsequent calls. This shouldn't be a problem if you call
Stats() every X seconds (>10) and quickly dump it to disk.